requirements.txt
astroid==2.0.4
Click==7.0
colorama==0.4.0
cycler==0.10.0
DateTime==4.3
Flask==1.0.2
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
matplotlib==3.0.1
mccabe==0.6.1
numpy==1.15.4
pandas==0.23.4
pylint==2.1.1
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
six==1.11.0
Werkzeug==0.14.1
wrapt==1.10.11
zope.interface==4.6.0
